requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone-client[grpc]
langchain-pinecone
langchain-openai
langchain-community
langchain-experimental
pyreadline3; sys_platform == 'win32'
-e .